\name{RMR.ICP-package}
\alias{RMR.ICP-package}
\alias{RMR.ICP}
\docType{package}
\title{
\packageTitle{RMR.ICP}
}
\description{
\packageDescription{RMR.ICP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RMR.ICP}
\packageIndices{RMR.ICP}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{RMR.ICP}

Maintainer: \packageMaintainer{RMR.ICP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one ~~
~~ per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
